requirements.dev.txt
add-trailing-comma
altair==4.1.0
aria2p==0.9.1
astroid==2.3.3
azure-identity
azure-mgmt-authorization
azure-mgmt-containerservice
azure-mgmt-resource
azure-mgmt-storage
azure-storage-file-share
black==22.3.0
certifi==2019.9.11
cryptography==36.0.1
cycler==0.10.0
Cython==0.29.14
deepdiff==5.7.0
docker
editorconfig-checker==2.4.0
flake8==4.0.1
flask-cors==3.0.10
flask==1.1.2
flask_cors==3.0.10
flask_socketio==5.2.0
flloat==0.3.0
geopy==2.0.0
guppy3==3.0.9
holidays==0.10.3
isort==4.3.21
jinja2==2.11.3
kiwisolver==1.1.0
kubernetes==21.7.0
lazy-object-proxy==1.4.3
markupsafe==2.0.1
matplotlib==3.5.2
mccabe==0.6.1
networkx==2.4
networkx==2.4
numpy<1.20.0
palettable==3.3.0
pandas==0.25.3
pre-commit==2.19.0
prompt_toolkit==2.0.10
psutil==5.8.0
ptvsd==4.3.2
pulp==2.6.0
pyaml==20.4.0
PyJWT==2.4.0
pyparsing==2.4.5
python-dateutil==2.8.1
PyYAML==5.4.1
pyzmq==19.0.2
recommonmark~=0.6.0
redis==3.5.3
requests==2.25.1
scipy==1.7.0
setuptools==58.0.4
six==1.13.0
sphinx==1.8.6
sphinx_rtd_theme==1.0.0
streamlit==0.69.1
stringcase==1.2.0
tabulate==0.8.5
termgraph==0.5.3
torch==1.6.0
torchsummary==1.5.1
tqdm==4.51.0
urllib3==1.26.5
wrapt==1.11.2
zmq==0.0.0
